# 07-rl-simulation-control-platform/contracts/http/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RL Simulation Control API",
    "version": "0.3.0",
    "description": "Public training-evidence reads and operator-authenticated control commands. Accepted commands are not execution confirmation."
  },
  "servers": [{"url": "/", "description": "Current application origin"}],
  "security": [{"operatorBearer": []}],
  "paths": {
    "/health/live": {"get": {"operationId": "live", "security": [], "responses": {"200": {"description": "Process is alive"}}}},
    "/health/ready": {"get": {"operationId": "ready", "security": [], "responses": {"200": {"description": "Dependencies are ready"}, "503": {"description": "Dependency unavailable", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}}},
    "/api/v1/environments": {"get": {"operationId": "listEnvironments", "security": [], "responses": {"200": {"description": "Registered environments", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Environment"}}}}}}}},
    "/api/v1/environments/{id}": {"get": {"operationId": "getEnvironment", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}], "responses": {"200": {"description": "Environment detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Environment"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/policies": {"get": {"operationId": "listPolicies", "security": [], "responses": {"200": {"description": "Allowlisted policies", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Policy"}}}}}}}},
    "/api/v1/policies/{id}": {"get": {"operationId": "getPolicy", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}], "responses": {"200": {"description": "Registered policy with its visual overlay", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Policy"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/operator/session": {"get": {"operationId": "getOperatorSession", "responses": {"200": {"description": "The supplied Bearer token is an active operator credential", "content": {"application/json": {"schema": {"type": "object", "required": ["role"], "properties": {"role": {"type": "string", "const": "operator"}}}}}}, "401": {"description": "Operator authentication required", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}}},
    "/api/v1/training-runs": {
      "get": {"operationId": "listTrainingRuns", "security": [], "parameters": [{"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/RunStatusFilter"}, {"$ref": "#/components/parameters/BeforeCursor"}], "responses": {"200": {"description": "Bounded run collection; X-Next-Cursor is emitted when another page may exist", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunCollection"}}}}}},
      "post": {"operationId": "createTrainingRun", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateRunRequest"}}}}, "responses": {"202": {"description": "Run queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrainingRun"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"description": "Idempotency conflict", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}}
    },
    "/api/v1/training-runs/{id}": {"get": {"operationId": "getTrainingRun", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}], "responses": {"200": {"description": "Run detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrainingRun"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/training-runs/{id}/pause": {"post": {"operationId": "pauseTrainingRun", "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"202": {"description": "Pause requested", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrainingRun"}}}}, "409": {"$ref": "#/components/responses/InvalidState"}}}},
    "/api/v1/training-runs/{id}/resume": {"post": {"operationId": "resumeTrainingRun", "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"202": {"description": "Resume requested", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrainingRun"}}}}, "409": {"$ref": "#/components/responses/InvalidState"}}}},
    "/api/v1/training-runs/{id}/cancel": {"post": {"operationId": "cancelTrainingRun", "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"202": {"description": "Cancel requested", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrainingRun"}}}}, "409": {"$ref": "#/components/responses/InvalidState"}}}},
    "/api/v1/training-runs/{id}/metrics": {"get": {"operationId": "listRunMetrics", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/MetricFilter"}, {"$ref": "#/components/parameters/AfterTimeCursor"}], "responses": {"200": {"description": "Sampled metrics; X-Next-Cursor is emitted when another page may exist", "content": {"application/json": {"schema": {"type": "array", "maxItems": 200, "items": {"$ref": "#/components/schemas/MetricSample"}}}}}}}},
    "/api/v1/training-runs/{id}/episodes": {"get": {"operationId": "listRunEpisodes", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/AfterEpisodeCursor"}], "responses": {"200": {"description": "Run episodes; X-Next-Cursor is emitted when another page may exist", "content": {"application/json": {"schema": {"type": "array", "maxItems": 100, "items": {"$ref": "#/components/schemas/Episode"}}}}}}}},
    "/api/v1/episodes/{id}": {"get": {"operationId": "getEpisode", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}], "responses": {"200": {"description": "Episode", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Episode"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/episodes/{id}/transitions": {"get": {"operationId": "listTransitions", "security": [], "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/AfterStepCursor"}], "responses": {"200": {"description": "Bounded ordered transitions; X-Next-Cursor is emitted when another page may exist", "content": {"application/json": {"schema": {"type": "array", "maxItems": 200, "items": {"$ref": "#/components/schemas/Transition"}}}}}}}},
    "/api/v1/episodes/{id}/feedback": {"post": {"operationId": "createEpisodeFeedback", "parameters": [{"$ref": "#/components/parameters/OpaqueId"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FeedbackRequest"}}}}, "responses": {"201": {"description": "Feedback recorded"}, "400": {"$ref": "#/components/responses/BadRequest"}}}},
    "/ws/v1/runs/{id}": {"get": {"operationId": "subscribeRun", "security": [], "description": "WebSocket upgrade authenticated by the rl-run-v1 subprotocol plus the operator token. Messages only notify clients to resync confirmed state over authenticated REST.", "x-websocket-auth": {"subprotocols": ["rl-run-v1", "operator-token"]}, "parameters": [{"$ref": "#/components/parameters/OpaqueId"}], "responses": {"101": {"description": "Switching protocols"}, "401": {"description": "Operator authentication required", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}}}
  },
  "components": {
    "securitySchemes": {
      "operatorBearer": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque operator token", "description": "Required only for commands and feedback. Public evidence reads need no token."}
    },
    "parameters": {
      "OpaqueId": {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 8, "maxLength": 100, "pattern": "^[A-Za-z0-9._:-]+$"}},
      "PageSize": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
      "RunStatusFilter": {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["draft", "queued", "running", "pausing", "paused", "cancelling", "cancelled", "succeeded", "failed"]}},
      "BeforeCursor": {"name": "before", "in": "query", "description": "Exclusive UTC created-at cursor from X-Next-Cursor.", "schema": {"type": "string", "format": "date-time"}},
      "MetricFilter": {"name": "metric", "in": "query", "schema": {"type": "string", "enum": ["episode_reward", "moving_average_reward", "epsilon", "loss", "episode_steps", "collisions", "success_rate", "action_up", "action_right", "action_down", "action_left", "training_examples", "prediction_error", "accumulated_error", "rollout_risk"]}},
      "AfterTimeCursor": {"name": "after", "in": "query", "description": "Exclusive UTC sampled-at cursor from X-Next-Cursor.", "schema": {"type": "string", "format": "date-time"}},
      "AfterEpisodeCursor": {"name": "afterEpisode", "in": "query", "description": "Exclusive episode-number cursor.", "schema": {"type": "integer", "minimum": 0}},
      "AfterStepCursor": {"name": "afterStep", "in": "query", "description": "Exclusive transition step cursor.", "schema": {"type": "integer", "minimum": 0}}
    },
    "responses": {
      "BadRequest": {"description": "Invalid request", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
      "NotFound": {"description": "Resource not found", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
      "InvalidState": {"description": "Invalid state transition", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
      "Coordinate": {"type": "object", "additionalProperties": false, "required": ["row", "column"], "properties": {"row": {"type": "integer", "minimum": 0, "maximum": 20}, "column": {"type": "integer", "minimum": 0, "maximum": 20}}},
      "Environment": {"type": "object", "additionalProperties": false, "required": ["id", "slug", "name", "version", "rows", "columns", "start", "goal", "obstacles", "rewardMap", "observationSpace", "actionSpace"], "properties": {"id": {"type": "string", "format": "uuid"}, "slug": {"type": "string"}, "name": {"type": "string"}, "version": {"type": "string"}, "rows": {"type": "integer"}, "columns": {"type": "integer"}, "start": {"$ref": "#/components/schemas/Coordinate"}, "goal": {"$ref": "#/components/schemas/Coordinate"}, "obstacles": {"type": "array", "items": {"$ref": "#/components/schemas/Coordinate"}}, "rewardMap": {"type": "object", "required": ["step", "collision", "goal"], "properties": {"step": {"type": "number"}, "collision": {"type": "number"}, "goal": {"type": "number"}}}, "observationSpace": {"type": "string", "const": "Discrete(36)"}, "actionSpace": {"type": "array", "prefixItems": [{"const": "up"}, {"const": "right"}, {"const": "down"}, {"const": "left"}], "items": false}}},
      "Policy": {"type": "object", "required": ["id", "algorithm", "version", "artifactUri", "sha256", "provenance", "observationSpace", "actionSpace", "actionByState", "enabled"], "properties": {"id": {"type": "string", "format": "uuid"}, "algorithm": {"type": "string", "enum": ["q-learning", "sarsa", "dqn", "world-model"]}, "version": {"type": "string"}, "artifactUri": {"type": "string", "pattern": "^artifact://"}, "sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "provenance": {"type": "string"}, "observationSpace": {"type": "string"}, "actionSpace": {"type": "array", "items": {"type": "string"}}, "actionByState": {"type": "object", "additionalProperties": {"type": "string", "enum": ["up", "right", "down", "left", "blocked"]}}, "enabled": {"type": "boolean"}}},
      "CreateRunRequest": {"type": "object", "additionalProperties": false, "required": ["environmentId", "policyId", "seed", "maxSteps"], "properties": {"environmentId": {"type": "string", "format": "uuid"}, "policyId": {"type": "string", "format": "uuid"}, "seed": {"type": "integer", "minimum": 0, "maximum": 2147483647}, "maxSteps": {"type": "integer", "minimum": 1, "maximum": 200}}},
      "TrainingRun": {"type": "object", "required": ["id", "environmentId", "policyId", "status", "attempt", "seed", "maxSteps", "createdAt", "updatedAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "environmentId": {"type": "string", "format": "uuid"}, "policyId": {"type": "string", "format": "uuid"}, "status": {"type": "string", "enum": ["draft", "queued", "running", "pausing", "paused", "cancelling", "cancelled", "succeeded", "failed"]}, "attempt": {"type": "integer", "minimum": 1}, "seed": {"type": "integer"}, "maxSteps": {"type": "integer"}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}, "terminalReason": {"type": ["string", "null"]}}},
      "RunCollection": {"type": "object", "required": ["items", "limit"], "properties": {"items": {"type": "array", "maxItems": 200, "items": {"$ref": "#/components/schemas/TrainingRun"}}, "limit": {"type": "integer"}}},
      "Episode": {"type": "object", "required": ["id", "runId", "episodeNumber", "status", "totalReward", "stepCount", "collisions", "terminalReason", "startedAt", "completedAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "runId": {"type": "string", "format": "uuid"}, "episodeNumber": {"type": "integer", "minimum": 1}, "status": {"type": "string", "enum": ["succeeded", "truncated", "cancelled", "failed"]}, "totalReward": {"type": "number"}, "stepCount": {"type": "integer"}, "collisions": {"type": "integer"}, "terminalReason": {"type": "string"}, "startedAt": {"type": "string", "format": "date-time"}, "completedAt": {"type": "string", "format": "date-time"}}},
      "Transition": {"type": "object", "required": ["id", "episodeId", "stepIndex", "state", "action", "nextState", "reward", "terminated", "truncated", "sampledAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "episodeId": {"type": "string", "format": "uuid"}, "stepIndex": {"type": "integer", "minimum": 0}, "state": {"$ref": "#/components/schemas/Coordinate"}, "action": {"type": "string", "enum": ["up", "right", "down", "left"]}, "nextState": {"$ref": "#/components/schemas/Coordinate"}, "reward": {"type": "number"}, "terminated": {"type": "boolean"}, "truncated": {"type": "boolean"}, "sampledAt": {"type": "string", "format": "date-time"}, "predictedState": {"$ref": "#/components/schemas/Coordinate"}, "predictedNextState": {"$ref": "#/components/schemas/Coordinate"}, "stepError": {"type": "number", "minimum": 0}, "accumulatedError": {"type": "number", "minimum": 0}, "modelVersion": {"type": "string"}}},
      "MetricSample": {"type": "object", "required": ["runId", "metric", "value", "unit", "step", "sampledAt"], "properties": {"runId": {"type": "string", "format": "uuid"}, "metric": {"type": "string", "enum": ["episode_reward", "moving_average_reward", "epsilon", "loss", "episode_steps", "collisions", "success_rate", "action_up", "action_right", "action_down", "action_left", "training_examples", "prediction_error", "accumulated_error", "rollout_risk"]}, "value": {"type": "number"}, "unit": {"type": "string"}, "step": {"type": "integer", "minimum": 0}, "sampledAt": {"type": "string", "format": "date-time"}}},
      "FeedbackRequest": {"type": "object", "additionalProperties": false, "required": ["category", "note"], "properties": {"category": {"type": "string", "enum": ["clear", "unexpected", "loop", "collision", "other"]}, "note": {"type": "string", "minLength": 1, "maxLength": 500}}},
      "Problem": {"type": "object", "required": ["type", "title", "status", "code", "traceId"], "properties": {"type": {"type": "string", "format": "uri-reference"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "code": {"type": "string"}, "traceId": {"type": "string"}}}
    }
  }
}